iscc: allow interruption of computation from the keyboard again
Support for interrupting the computation was added in e2d8fba (iscc:
allow interruption of computation from the keyboard,
Thu Mar 17 10:07:22 2011 +0100), but broken in 96c7209 (iscc.c:
use isl_stream_is_empty instead of reading eof field of isl_stream,
Wed Dec 17 17:30:25 2014 +0100) because the subsequent call
to isl_stream_is_empty would think the stream is empty because
no token could be constructed.

Move the detection of interruption to the main loop, also handling
the case of interruption during input and combining it with the call
to isl_ctx_resume.
Handling of interruption during input is not ideal because the first
token after the interrupt is dropped.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>