barvinok_enumerate_e: drop option to use Parker's DFA based counting
In fact, since this is the only remaining use
of Parker's library, remove it completely.

This was only ever meant as a means to compare
different ways of counting and it is not clear
if it still compiles.
In particular, it (also) uses the Omega library and
use of that library will be completely removed in the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>