scale.c: Param_Polyhedron_Domain: only set Facets in result if set in input
A Param_Vertices of a Param_Polyhedron created by ISL_P2PP
does not have its Facets set.  Since the user may set the chambers option
to BV_CHAMBERS_ISL, Param_Polyhedron_Domain should not assume
that this field has been set.

Simply refrain from setting the field in the result
if it was not set in the input.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>