series.cc: enumerate_union_series: drop skew_into_positive_orthant
This function seems to try to move the non-parametric variables
into the positive orthant.  However, since these dimensions should
be bounded, there should be no need for them to be non-negative.
Moreover, the code does not look like it does the skewing correctly.
In particular, the loop over the variables that are supposed
to be made positive considered a number of variables
equal to the number of parameters instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>