barvinok_enumerate_e: drop Omega simplification
This has not been tested in a while and it is not even clear
which version of Omega this is supposed to use.
Furthermore, the Omega library no longer appears to be maintained.

If a user really needs such simplifications, they will just
have to do them separately.
The iscc interface also provides similar simplifications through isl.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>