Polyhedron_Lattice_Width_Directions: do not skip duplicate directions
For non-parametric polytopes, a given direction always yields the same
width, but this is not the case for parametric polytopes.
For example, the new test case is flattest in the (1,0) direction,
with width n/2 or 10, depending on the value of the parameter n.
If the width in the (1,0) direction is only computed once,
then only one of those widths is found and the other is replaced
by the width n found in other directions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>