gen_fun::summate: properly keep track of number of attempts
Since 95ea4d5 (randomvector: double size of coefficients on each attempt,
Wed Mar 10 15:11:58 2010 +0100), the maximum value of the elements
in the random vector is increased on every attempt, but only
np_base::start would actually keep track of the number of attempts.
gen_fun::summate, on the other hand, would just reset the number
of attempts to zero on every iteration, defeating the purpose
of keeping track of the number of attempts.
Move the initialization out of the loop.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>