barvinok.cc: basic_map_card: explicitly reintroduce domain
The function basic_map_card turns input dimensions into parameters
in the input and then turns those parameters back into input dimensions
in the result.  The input tuple is completely removed in the input,
but it was not getting explicitly introduced again in the output.
Instead, it was relying on isl_pw_qpolynomial_move_dims implicitly
introducing the input tuple, but this is an undocumented feature
that is not guaranteed to persist.
Introduce the domain explicitly.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>