distribute both versions of interface/isl.py.top
This was missing from f24cadb22a (introduce special version of isl.py.top
for use with pet, Tue Dec 8 15:58:18 2015 +0100).
Since a distribution always contains the pet submodule,
only packaging the pet version of interface/isl.py.top
should be fairly harmless, but it does result in an error message
about the missing file and it may break if the user decides
not to use the pet submodule.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>