configure.ac: override --with-isl option for submodules when set to bundled
If the user explicitly passes the --with-isl=bundled option to configure,
then it will be passed down to submodules along with
the --with-isl-builddir=../isl option that gets added
to let submodules use the isl that is bundled with the top-level project.
The submodule complains about these conflicting options.
Override the --with-isl=bundled with an additional --with-isl=build option.

Note that a user shouldn't normally need to set the --with-isl=bundled
because that is the default, but it should work nonetheless.

Reported-by: Kornilios Kourtis <kkourt@kkourt.io>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>