barvinok_bound: fix --iterate option
In particular, adjust the code to the changed isl_pw_qpolynomial space.
This was missing from barvinok-0.34-25-g235972b7ef (update isl
for change in space of isl_pw_* objects, Sun Sep 4 15:04:00 2011 +0200).
Add a test case for this option to ensure it does not get broken again.
Note that --iterate only works for inputs without parameters.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>