configure.ac: require cddlib 0.94k by checking for dd_SetMatrixObjective
Version 0.94k of cddlib installs headers in a cddlib/ subdirectory.
Adjust the locations where cddlib is used and check
for a function (dd_SetMatrixObjective) that is new to 0.94k
to ensure the cddlib version found did not just happen to have
its headers appear in a cddlib/ subdirectory, as apparently
can happen in some distributions.
Since cddlib 0.94k is now required, the check for cddlib 0.94e
is no longer needed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>