ISL_P2PP: store (potentially) updated constraints in Param_Polyhedron
The constraints used by isl_basic_set_compute_vertices may be
different from those in the original Polyhedron because isl
may have performed some simplifications on them.
This is especially the case since isl commit isl-0.26-139-g6dff4a1e27
(simplify constraints by checking for residues,
Mon Jul 12 14:34:53 2021 +0200), which was brought in by
barvinok-0.41.7-11-gc340b56cc4 (update isl for support for recent versions
of clang, Sun Jul 7 15:12:41 2024 +0200).
The constraints of the Param_Polyhedron are used to construct
the vertex cones, so it is important for them to match those
that were used to compute the vertices.

An alternative would be to construct rational isl sets
to avoid any such simplifications, but there is currently
no proper interface for constructing such sets.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>