scale: use scaled Param_Polyhedron constraints
Instead of scaling both the Param_Polyhedron and the Polyhedron
from which it was derived, only scale the Param_Polyhedron and
derive a scaled Polyhedron from the constraints of the result.
The constraints of the Param_Polyhedron may have been further
simplified and the scaling is based on the corresponding vertices,
so it is important to scale the (possibly) simplified constraints.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>