scale.c: Param_Polyhedron_Scale_Integer_Slow: remove dependence on P parameter
If P is specified, then its constraints get scaled as well,
but otherwise this function does not depend on P,
except that it is also used to determine the number of variables.

Extract this number from the Param_Polyhedron,
as is done in Param_Polyhedron_Scale_Integer_Fast.

This allows the next commit to pass a NULL to Param_Polyhedron_Scale.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>