scale.c: PP_enumerate_narrow_flated: use scaled Param_Polyhedron constraints
As in the previous commit, only scale the Param_Polyhedron and
derive a scaled Polyhedron (if needed) from the constraints of the result
to take into account a possible simplification of the constraints.

Note that there is no strict need to call Param_Polyhedron2Polyhedron here
since the Param_Polyhedron will be freed anyway.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>